Integrated the new replay-export code into the game.
This is alpha version code now. It's not yet ready for general use.

Now it's possible to export a replay directly to a webm file from the game.

The resolution is currently hardcoded to 800x600, and the bit depth is hardcoded to 24 bpp.

I have only tested this on a 32 bit Linux system so far.

The webm writer currently keeps all encoded frames in memory until the end of the recording, so don't expect to be able to record gigabytes of video with it just yet.

I will add more fixes to this code later, and hopefully it will become usable soon.